+++ sim/encounters/naxxramas/patchwerk25_ai.go
@@ -65,7 +65,7 @@ func (ai *Patchwerk25AI) Reset(*core.Simulation) {
 }
 
 func (ai *Patchwerk25AI) registerHatefulStrikeSpell(target *core.Target) {
-	actionID := core.ActionID{SpellID: 59192}
+	actionID := core.ActionID{SpellID: 28308}
 
 	ai.HatefulStrike = target.RegisterSpell(core.SpellConfig{
 		ActionID:    actionID,